+++ src/ducks/reducer.js
@@ -3,7 +3,7 @@ import auth from './auth';
 import programs from './programs';
 import classrooms from './classrooms';
 import assignments from './assignments';
-import mapexplorer from './wildcam-mapexplorer';
+import mapexplorer from '../modules/wildcam-maps/ducks/index';
 import notification from './notification';
 import caesarExports from './caesar-exports';
