+++ src/lib/zooniversal-translator.es.js
@@ -163,7 +163,7 @@ let SpanishTranslations = {
 };
 
 //HACK: database support for WildCam Darien.
-import mapConfig from './wildcam-darien.mapConfig.js';
+import mapConfig from '../programs/darien/wildcam-darien.map-config.js';
 const databaseAdditions = {};
 mapConfig.map.filters.data_choice.options.map(item => {
   const databaseValue = item.value;
