+++ src/modules/wildcam-maps/components/CameraViewer.jsx
@@ -13,7 +13,7 @@ import CameraViewerPaging from './CameraViewerPaging';
 import {
   MAPEXPLORER_INITIAL_STATE, MAPEXPLORER_PROPTYPES,
   MAPEXPLORER_CAMERA_STATUS,
-} from '../../../ducks/wildcam-mapexplorer';
+} from '../ducks/index.js';
 
 const ITEMS_PER_PAGE = 6;
