+++ src/modules/wildcam-maps/components/CameraViewerMetadata.jsx
@@ -9,7 +9,7 @@ import { ZooTran, ZooTranExists } from '../../../lib/zooniversal-translator.js';
 import {
   MAPEXPLORER_INITIAL_STATE, MAPEXPLORER_PROPTYPES,
   MAPEXPLORER_CAMERA_STATUS,
-} from '../../../ducks/wildcam-mapexplorer';
+} from '../ducks/index.js';
 
 const CameraViewerMetadata = (props) => {
   if (props.activeCameraMetadataStatus === MAPEXPLORER_CAMERA_STATUS.FETCHING) {
