+++ src/modules/wildcam-maps/components/CameraViewerPaging.jsx
@@ -12,7 +12,7 @@ import { ZooTran } from '../../../lib/zooniversal-translator.js';
 import {
   MAPEXPLORER_INITIAL_STATE, MAPEXPLORER_PROPTYPES,
   MAPEXPLORER_CAMERA_STATUS,
-} from '../../../ducks/wildcam-mapexplorer';
+} from '../ducks/index.js';
 
 const CameraViewerData = (props) => {
   if (props.activeCameraDataStatus === MAPEXPLORER_CAMERA_STATUS.SUCCESS &&
