+++ src/modules/wildcam-maps/containers/MapExplorer.jsx
@@ -8,17 +8,17 @@ This feature is a 'data visualisation tool' allows Teachers and Students to view
 aggregated data from a Zooniverse project on a visual, geographical map.
 
 Requires:
-* a project-specific `mapConfig` object.
-* (External dependency) an external database containing the map data for said
+- a project-specific `mapConfig` object.
+- (External dependency) an external database containing the map data for said
   project, which can be queried with SQL SELECT commands. For example, Carto.
   
 Notable components:
-* MapVisuals: handles the visual display of the map.
-* MapControls: allows users to select the kind of data they want.
-* ducks/wildcam-mapexplorer: stores the map explorer data (e.g. view state) AND
+- MapVisuals: handles the visual display of the map.
+- MapControls: allows users to select the kind of data they want.
+- ducks/index.js: stores the map explorer data (e.g. view state) AND
     connects to the external database.
-* lib/mapexplorer-helper: handles general utility features, such as constructing
-    SQL "where" clauses from selected filters.
+- lib/wildcam-maps-helpers: handles general utility features, such as
+    constructing SQL "where" clauses from selected filters.
 
 ********************************************************************************
  */
@@ -39,7 +39,7 @@ import {
 
 import {
   MAPEXPLORER_INITIAL_STATE, MAPEXPLORER_PROPTYPES,
-} from '../../../ducks/wildcam-mapexplorer';
+} from '../ducks/index.js';
 
 class MapExplorer extends React.Component {
   constructor(props) {
