+++ src/modules/wildcam-maps/lib/wildcam-maps-helpers.js
