+++ src/programs/darien/wildcam-darien.map-config.js
